Refactor utils into pull_request.py
Using the utils/<file>.py setup wasn't the right approach. Moving those functions
into a PR-specific files, we can enhance this further down the road as there are
more functions and files added.
